Make multi service a trait in preparation for caching
